requirements.txt
flask==3.0.0
requests==2.31.0
gspread==6.1.0
google-auth==2.25.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pandas==2.1.4
python-dotenv==1.0.0
